refactor: move /agents view + orphans into internal/server/fleet
Phase 3c of the webhook split (issue #250). The fleet snapshot view
served at GET /agents and the orphan-detection cache that feeds /status
+ /agents/orphans/status now live alongside the CRUD surface in
internal/server/fleet, completing the fleet domain.

Architecture:
  - fleet.Handler gains an orphan cache (orphanMu/orphanCache) plus
    RefreshOrphansFromCfg, RefreshOrphansFromDB, and OrphansSnapshot
    methods, and a HandleOrphansStatus HTTP handler.
  - fleet.Handler grows a HandleAgentsView method (was webhook's
    handleAPIAgents) reading cfg + StatusProvider + RuntimeStateProvider.
  - The handler is now constructed in webhook.NewServer (cfg-only mode)
    so it works for tests that never call WithStore. WithStore wires
    the database via a new SetDB method; until then CRUD routes are
    skipped at registration time.
  - WithRuntimeState forwards into the handler via SetRuntimeState so
    the fleet view reports running/idle status.
  - webhook.Server's /status path and reloadCron now go through the
    handler (OrphansSnapshot, RefreshOrphansFromDB, RefreshOrphansFromCfg)
    instead of the deleted webhook-internal helpers.
  - webhook/orphans.go is gone; the agents view + types in webhook/api.go
    moved to fleet/view.go (api.go now only owns /config until Phase 4).

Tests:
  - TestCanonicalModels and TestComputeOrphanedAgents move to fleet
    (testing private helpers).
  - TestAgentsOrphansEndpointAndStatusSummary stays in webhook,
    importing the now-public fleet.OrphanedAgentsSnapshot type and
    exercising the routes via buildHandler.
  - api_test.go now hits the router via buildHandler() and decodes
    into local mirror types (view*JSON in crud_wire_test.go) so the
    tests don't depend on fleet's unexported wire shapes.

webhook drops ~265 lines net. The webhook package no longer owns any
fleet-domain logic — only HMAC + GitHub event parsing, /config,
/export, /import, /status, /run, and the proxy mount remain. Phase 4
will pull /config + /export + /import into internal/server/config.
